# Optimize `read_sql` + `head`