--- a/mars/dataframe/datasource/read_sql.py
+++ b/mars/dataframe/datasource/read_sql.py
@@ -56,13 +56,14 @@ class DataFrameReadSQL(DataFrameOperand, DataFrameOperandMixin):
     _high_limit = AnyField('high_limit')
     _left_end = BoolField('left_end')
     _right_end = BoolField('right_end')
+    _nrows = Int64Field('nrows')
 
     def __init__(self, table_or_sql=None, selectable=None, con=None, schema=None,
                  index_col=None, coerce_float=None, parse_dates=None, columns=None,
                  engine_kwargs=None, row_memory_usage=None, method=None,
                  incremental_index=None, offset=None, partition_col=None,
                  num_partitions=None, low_limit=None, high_limit=None, left_end=None,
-                 right_end=None, output_types=None, gpu=None, **kw):
+                 right_end=None, nrows=None, output_types=None, gpu=None, **kw):
         super().__init__(_table_or_sql=table_or_sql, _selectable=selectable, _con=con,
                          _schema=schema, _index_col=index_col, _coerce_float=coerce_float,
                          _parse_dates=parse_dates, _columns=columns,
@@ -70,7 +71,8 @@ class DataFrameReadSQL(DataFrameOperand, DataFrameOperandMixin):
                          _method=method, _incremental_index=incremental_index, _offset=offset,
                          _partition_col=partition_col, _num_partitions=num_partitions,
                          _low_limit=low_limit, _left_end=left_end, _right_end=right_end,
-                         _high_limit=high_limit, _output_types=output_types, _gpu=gpu, **kw)
+                         _high_limit=high_limit, _nrows=nrows, _output_types=output_types,
+                         _gpu=gpu, **kw)
         if not self.output_types:
             self._output_types = [OutputType.dataframe]
 
@@ -149,6 +151,10 @@ class DataFrameReadSQL(DataFrameOperand, DataFrameOperandMixin):
     @property
     def right_end(self):
         return self._right_end
+
+    @property
+    def nrows(self):
+        return self._nrows
 
     def _get_selectable(self, engine_or_conn, columns=None):
         import sqlalchemy as sa
@@ -435,11 +441,17 @@ class DataFrameReadSQL(DataFrameOperand, DataFrameOperandMixin):
                 if op.offset > 0:
                     query = query.offset(op.offset)
 
+            if op.nrows is not None:
+                query = query.limit(op.nrows)
+
             df = pd.read_sql(query, engine, index_col=op.index_col,
                              coerce_float=op.coerce_float,
                              parse_dates=op.parse_dates)
             if op.method == 'offset' and op.index_col is None and op.offset > 0:
-                df.index = pd.RangeIndex(op.offset, op.offset + out.shape[0])
+                index = pd.RangeIndex(op.offset, op.offset + out.shape[0])
+                if op.nrows is not None:
+                    index = index[:op.nrows]
+                df.index = index
             ctx[out.key] = df
         finally:
             engine.dispose()