<!--
Thank you for your contribution!

Please review https://github.com/mars-project/mars/blob/master/CONTRIBUTING.rst before opening a pull request.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

This PR optimized `read_sql` and `head`. 

What's more, if input DataFrame's index is RangeIndex started with 0, `df.loc[:n]` will be rewritten with `df.iloc[:n]`, so that it can leverage the universal optimization about head, and ensure correctness when input DataFrame does not ensure index incremental.

|   | Master | This PR |
| ------------- | ------------- | ------------- |
| Memory Usage | 191.38M  | 61.73M |
| Time  | 12.2s | 9.5s |

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this change? -->

NA